ReactElementOrText no longer to [External]
Bridge would error if you tried to declare a List<ReactElementOrText>
when ReactElementOrText was [External] and did not map on to an existing
JavaScript class. Unlike the ReactElement class (which has a React
counterpart, even though it's private and so not tied to the
ReactElement class with a [Name] attribute in case it changes in the
future) the ReactElementOrText has no corresponding React class and so
it does not need to be external. This change makes the creation of
dynamic children elements much easier.