Workaround for PureComponent issue with Props types with struct properties
PureComponent does not work effectively for Props that have properties
that are structs due to how these fields are initialised in Bridge. This
update includes a workaround if the Props types are defined using the
IAmImmutable helpers in the ProductiveRage.Immutable library (require
version 1.6.7). This workaround should only be required temporarily,
hopefully it will be addressed in Bridge 1.12.