Tweaks to sample app to prevent the React warning - Each child in an array or iterator should have a unique "key" prop
Two changes were required. The dynamic elements (the individual TODO
cards) need to have a Key value on them, which is easy enough since each
Todo instance has an Id property (though I changed this to always
increment, rather than taking the count of the Todo list, to ensure that
values are not reused when items are removed and then new ones added).
The non-dynamic elements (the header and the input form) are now
specified using the original method signature, which is not intended for
use with dynamic elements and so the warning is avoided (there is
potential for a naughty workaround to avoid the dynamic children warning
by passing a pre-populated array reference instead of listing the static
elements one-by-one as params - I'd like to look into adding an analyser
to detect this).