Changed new ReactStyle class to be like the other React attribute classes; write-only
This doesn't stop the MergeWith method working or shared style instances
from being defined (as is done in the TODO sample application), it is
only intended to emphasise that component attribute classes should have
properties set before use and then not changed. The downside is that it
means that the properties from the Bridge CSSStyleDeclaration class need
to be repeated in the new ReactStyle class but that doesn't inflate the
runtime library at all since it's an [ObjectLiteral] class - and this
change means that properties no longer need to be created "new" in the
derived ReactStyle class, they may be declared initially as "Any<string,
int>" rather than being a "string" in the CSSStyleDeclaration and
"Any<string, int>" in the ReactStyle. The final benefit is that the
methods available on the CSSStyleDeclaration class (GetEnumerator,
GetPropertyValue, RemoveProperty, etc..) are no longer present, which
means that code won't be written that relies upon them - which is good
because they wouldn't have been present at run time since ReactStyle is
[ObjectLiteral].