Added alternative component creation mechanism that enables greater performance for huge numbers of elements
The StaticComponent class may be used to generate React stateless
elements (either pure - meaning that their content varies only by their
props value and that the props values support shallow comparisons - or
just stateless - meaning that their content may vary by other data or
that their props type does not support shallow comparison). Since static
functions are used for rendering instead of class instances, there is
less overhead in the rendering, which may be important for pages that
have to support updating 1000s of elements. In the future, the
PureComponent and StatelessComponent classes *may* be marked as
obsolete, with a recommendation to use static function variations
instead.