Extended bound method comparison for the PureComponent so that more re-renders may be avoided
If two Props references are of the same type and have properties that
are all the same references then a PureComponent does not need to
re-render - that's the easiest case. Sometimes there are bound methods
are Props properties and these won't be the same reference, but they
will be function references with $scope and $method properties that are
consistent. Sometimes even this won't catch it (if the method is an
anonymous function) so another fallback has been added that will confirm
that the two $scope values are the same and then that the ToString()
representation of the methods are the same.