Added ComponentDidMount and ComponentDidUpdate to the PureComponent<T> base class (bumped version number for release)
It's not possible to add these to StatelessComponent<T> since those
components are created as render functions, they don't have a full
backing React component class and so can't have lifecycle methods. It's
also not possible with the library as it stands now to hook these events
when using the StaticComponent.Pure method - it's not technically
impossible to make it work but there are complications around how to
cache the component class (this may be revisited in the future if
there's any demand).