Added ComponentWillUnmount support to the PureComponent (it already has ComponentDidMount and ComponentDidUpdate)
Access to the mounting life cycle in this way makes it easier to create
components that need to access the DOM node using the Refs callback
(they need to be able to throw away the DOM node reference when the
component is unmounted).

Note: These can't be added to the StatelessComponent since that uses the
pure-function-as-component functionality under the hood whereas the
PureComponent is essentially a "full" component (so that it can
implement ShouldComponentUpdate).