Rewrote React component initialisation for compatibility with 16.0
React 16.0 won't support React.createClass so need to bring to
initialise components in a similar way to ES6 class-based components do.
A side effect of this is that PureComponent will support the full set of
lifecycle methods as Component does, it seems unlikely that
optimisations will appear that enable components to support
shouldComponentUpdate without also supporting the full lifecycle.